requirements.txt
google-genai
python-dateutil
python-dotenv
